cockpit.js: split helper functions to new file
Creates a file called common.js which contains various self-contained
helper functions which have many users in cockpit.js.

This is a clean move.  Outside of adding `export` on each `function`,
and `import` in `cockpit.js`, nothing else changed.
